Create swsu.txt
swsu.org and swsu .ru are the official sites of South-West State University. Here the links where you can see that it are belong to the same university www.swsu.ru/res-inet. The footer swsu.org site you can find the contacts "cido@swsu.org". Access to e-mail via the subdomain doc.swsu.org, which is a subdomain of the university website. For example, my mail is star@swsu.org
KSTU is the old name of the same university, it was renamed in 2010, here's the proof en.wikipedia.org/wiki/South-West_State_University  
Thank you!