punjab agricultural university
Hi there, As a PAU teacher i would like to add my university into your list. So that we can use the Visual Studio with Resharper which is wuite good tool for development. Our primary focus to developer the application for out University as well as to teach our students for their great  Currier.