create scdl domain
http://www.scdl.net/

http://www.scdl.net/distance-learning-in-information-technology.aspx
The IT industry has emerged as one of the fastest growing sectors of the Indian economy. Given the unprecedented growth of IT industry, organisations require manpower that is proficient in both information technology as well as business aspect. A successful IT enterprise will have the right mix of sterling technology systems and proven business strategies. Developing these systems requires extensive communication between technical developers, business managers and users.

With the rapid growth in international trade in various countries, Post Graduate Diploma in IT Management program with an emphasis on utilising information technologies for productivity and competitiveness, prepares you to succeed in your IT management career. The program's unique focus on the management of technology effectively prepares students for successful careers in the dynamic and fast-paced information technology marketplace.

In addition, the unique blend of courses will enable you to gain real world experience that will be invaluable to the development of your skills and your future employability. The curriculum produces students to succeed in IT management positions, which are projected to be amongst the fastest growing through the next decade.

Career Opportunities: 
Project Manager, IT Manager, System Analyst, Business Analyst, Technical Leaders, Software Consultant etc.