Add Huazhong University of Science and Technology
About University official website URL in english:
 http://english.hust.edu.cn/
IT-related course is offered by the university. pdf which can be downloaded in this webpage can give detailed proof: 
http://cs.hust.edu.cn/info/1076/1194.htm
Picture can prove the committed email is belong to this school. The picture is a PrtSc from my laptop, and I uploaded it to this website: 
https://postimg.cc/ctYVCzxj