Please add @estudante.sc.senai.br domain
My university is using old domains for students that already did a course long ago (which is my case), this old domain is not being accepted, only the new one is, the problem is that they can't update my student e-mail to the new domain. Please add the domain in the commit message so I can get the students licence as well. Thank you!