Add TOBB University of Economics and Technology domain
This pull request adds the email domain for TOBB University of Economics and Technology (etu.edu.tr) to the repository. 

TOBB University is a recognized educational institution offering IT-related courses and programs. The domain information has been added in accordance with the repository structure and guidelines.

Official website: https://www.etu.edu.tr