Add extension aof.anadolu.tr
Institution Name:
Anadolu University Open Education

Official Website URL:
https://www.anadolu.edu.tr/acikogretim
https://aosogrenci.anadolu.edu.tr/
https://aosdestek.anadolu.edu.tr

Distance education is a contemporary practice that “enables students to learn by utilizing communication technologies”. Distance education, which is practiced in many developed and developing countries, is a system in which people of all ages, incomes and occupational groups can continue their education by adjusting their learning according to their own speed and capacity without losing their productivity. Distance education, which includes education models that bring together students, instructors and educational tools in different places with communication technologies, is an educational philosophy that enables the use of the technology of the age. 
Anadolu University Open and Distance Education System is the first institution in our country to provide higher education with the distance education model in the modern sense.
Articles 5 and 12 of Law No. 2547, which entered into force on November 6, 1981 and reorganized Turkish higher education, granted Turkish universities the right to “Provide Continuous and Open Education”. Later, this task was assigned to Anadolu University, which had completed its scientific and technological deficiencies in a short time, with the Decree Law No. 41 issued on July 20, 1982. The Faculty of Open Education, which was born within the existing Faculty of Communication Sciences, was thus assigned to provide distance education services at the national level.
Starting the 1982-1983 academic year with two distance education programs in Economics and Business Administration, the Open Education System has continuously expanded its service areas. In 1993, the Open Education System was restructured according to the educational requirements of the day with the Decree Law No. 496 and the Economics and Business Administration Programs were transformed into four-year faculties. The Faculty of Open Education was assigned to carry out the practical work related to open education and to carry out associate degree, bachelor's degree completion and all kinds of certificate programs.
In addition to utilizing printed course materials and television broadcasts, the Faculty of Open Education uses advanced technologies such as the Internet, computer assisted instruction and videoconferencing to ensure the adaptation of students to these technologies and to create technological interaction. In addition, academic counseling and application services (face-to-face education) are also provided in appropriate areas.